test: 토큰 재발급 테스트 코드
