+++ src/main/java/com/hufs_cheongwon/service/UsersService.java
@@ -23,7 +23,7 @@ import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 public class UsersService {
 
     private final UsersRepository usersRepository;
@@ -32,7 +32,6 @@ public class UsersService {
     private String univCertKey;
     private String univName = "한국외국어대학교";
 
-    @Transactional
     public SignupResponse registerUser(LoginRequest request) throws IOException{
 
         String email = request.getEmail();
