+++ src/main/java/com/hufs_cheongwon/web/controller/UserController.java
@@ -1,10 +1,12 @@
 package com.hufs_cheongwon.web.controller;
 
 import com.hufs_cheongwon.common.Constant;
+import com.hufs_cheongwon.common.exception.UserNotFoundException;
 import com.hufs_cheongwon.common.security.JwtUtil;
 import com.hufs_cheongwon.service.RefreshTokenService;
 import com.hufs_cheongwon.service.UsersService;
 import com.hufs_cheongwon.web.apiResponse.ApiResponse;
+import com.hufs_cheongwon.web.apiResponse.error.ErrorStatus;
 import com.hufs_cheongwon.web.apiResponse.success.SuccessStatus;
 import com.hufs_cheongwon.web.dto.*;
 import jakarta.servlet.http.HttpServletResponse;
@@ -63,10 +65,14 @@ public class UserController {
     public ApiResponse<LoginResponse> reissueUserToken(
             @CookieValue(name = "refresh_token") String refreshToken, HttpServletResponse response) throws IOException{
 
+        if (refreshToken == null) {
+            throw new UserNotFoundException(ErrorStatus.COOKIE_EMPTY);
+        }
+
         LoginResponse reissueResponse = refreshTokenService.reissueToken(refreshToken);
 
         // Refresh Token을 쿠키에 설정
-        ResponseCookie refreshTokenCookie = ResponseCookie.from("refresh_token", refreshToken)
+        ResponseCookie refreshTokenCookie = ResponseCookie.from("refresh_token", reissueResponse.getTokenDto().getRefreshToken())
                 .httpOnly(true)
                 .secure(true)
                 .sameSite("Strict")
